anolis: KVM: ARM: Support KVM being compiled as a kernel module
ANBZ: #4924

This patch adds support for KVM ARM64 to be compiled as a kernel module.
It makes the CONFIG_KVM_ARM_HOST as a tristate option and adds a new
config option CONFIG_KVM_ARM_HOST_VHE_ONLY to ensure that kernel module
feature only supports for VHE system.

[ Merged commit("ARM64: KVM: Fix kvm running on non-vhe platform.") ]

+++ arch/arm64/kvm/arch_timer.c
@@ -1078,10 +1078,12 @@ void kvm_timer_hyp_uninit(void)
 
 	cpuhp_remove_state(CPUHP_AP_KVM_ARM_TIMER_STARTING);
 	if (info->physical_irq > 0) {
-		on_each_cpu((smp_call_func_t)disable_percpu_irq, (void *)host_ptimer_irq, 1);
+		on_each_cpu((smp_call_func_t)disable_percpu_irq,
+			    (void *)(long)host_ptimer_irq, 1);
 		free_percpu_irq(host_ptimer_irq, kvm_get_running_vcpus());
 	}
-	on_each_cpu((smp_call_func_t)disable_percpu_irq, (void *)host_vtimer_irq, 1);
+	on_each_cpu((smp_call_func_t)disable_percpu_irq,
+		    (void *)(long)host_vtimer_irq, 1);
 	free_percpu_irq(host_vtimer_irq, kvm_get_running_vcpus());
 }
